Connor Kelley <kelleyconnor@outlook.com>

+++ Homework/Assignment 1/Gaddis_8thEd_Chap2_Prob18/main.cpp
@@ -29,12 +29,13 @@ int main(int argc, char** argv) {
     //Process / Calculate
     nEDrnks = nCust/static_cast<float>(CNVPCT)*pEDrnks+0.5;     //Round Up for sake of whole person
     nCDrnks = nEDrnks/static_cast<float>(CNVPCT)*pCDrnks+0.5;   //Round Up for sake of whole person
+    
     //Output Everything
     cout<<"Energy Drinks Customer Survey Problem"<<endl;
     cout<<"Known/Given Values: "<<endl;
     cout<<"Number of customers = "<<nCust<<endl;
-    cout<<"Percentage Energy Drinkers = "<<static_cast<int>(pEDrnks)<<"%"<<endl;
-    cout<<"Percentage Citrus Drinkers/Energy Drinkers = "<<static_cast<float>(pCDrnks)<<"%"<<endl;
+    cout<<"Percentage Energy Drinkers = "<<static_cast<short>(pEDrnks)<<"%"<<endl;
+    cout<<"Percentage Citrus Drinkers/Energy Drinkers = "<<static_cast<short>(pCDrnks)<<"%"<<endl;
     cout<<"Calculated Values: "<<endl;
     cout<<"Number of Energy Drinkers = "<<nEDrnks<<endl;
     cout<<"Number of Citrus Drinkers = "<<nCDrnks<<endl;
